BUG#: 5902
TITLE: Compiler Warnings

DESCRIPTION:  Fix for type mismatch in assignment

+++ src/Pegasus/Common/CommonUTF.cpp
@@ -27,9 +27,9 @@
 // ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
-//==============================================================================
+//=============================================================================
 //
-//%/////////////////////////////////////////////////////////////////////////////
+//%////////////////////////////////////////////////////////////////////////////
 
 #include <Pegasus/Common/Config.h>
 #include <Pegasus/Common/Array.h>
@@ -162,7 +162,7 @@ int UTF16toUTF8(const Uint16** srcHead,
                 break;
             }
 
-            *tgt++ = *src++;
+            *tgt++ = (Uint8)*src++;
             continue;
         }
